move pinmfs algo in separate file
